Ledger cronjob (#28206)
* sync retry ledger

* add ledger fetch by transactor api

* sync retry ledger

* remove retry sleep

* ledger fetchby transactor add

* ledger fetchby transactor add

* ledger status queue config

* ledger status queue config

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* ledger async retry wip

* force async retry - debug

* formatting

* ledger async retry formatting

* ledger async retry job formatting

* ledger adj and bt core formatting

* async retry formatting

* clone payout change for payout failed for ledger

* async retry formatting

* async retry formatting

* async retry formatting

* async retry formatting

* async retry formatting

* ledger reverse shadow fav processed reversed fixes

* exception handling after async job dispatch

* exception handling after async job dispatch

* missed import fix

* async retry logic change | error handling | formatting

* formatting

* Added cron route for ledger cron jobs

* Refracted code

* Refracted code

* Added testing changes

* formatting

* Added testing changes

* tests fix

* formatting

* formatting

* skip fts calls for payout and fav in case of sync ledger call failures

* Added changes

* Added repo changes

* formatting

* Added more logs

* Added more logs

* Refracted code

* dummy for bvt

* dummy for bvt

* dummy for bvt

* Refracted code

* Resolved conflicts

* Added balance join query

* Added balance join query in reversals

* Reversals fix

* Reversals fix

* Reversals fix

* Reversals fix

* Fixed tests

* Resolved comments

* Resolved comments

* Resolved comments

* Resolved comments

* Fixed account create on event idempotency

* Resolved comments and added tests

+++ app/Jobs/LedgerStatus.php
@@ -31,6 +31,7 @@ class LedgerStatus extends Job
 
     protected $ledgerRequest;
     protected $feeSplit;
+    protected $retryEnabled;
     protected $transactorId;
     protected $transactorEvent;
 
@@ -56,12 +57,13 @@ class LedgerStatus extends Job
     const LEDGER_RESPONSE_MSG       = 'msg';
     const LEDGER_RECORD_NOT_FOUND   = 'record_not_found';
 
-    public function __construct(string $mode, array $ledgerRequest, array $feeSplit = null)
+    public function __construct(string $mode, array $ledgerRequest, array $feeSplit = null, bool $retryEnabled = true)
     {
         parent::__construct($mode);
 
-        $this->ledgerRequest    = $ledgerRequest;
-        $this->feeSplit         = $feeSplit;
+        $this->ledgerRequest = $ledgerRequest;
+        $this->feeSplit      = $feeSplit;
+        $this->retryEnabled  = $retryEnabled;
     }
 
     public function handle()
@@ -109,8 +111,12 @@ class LedgerStatus extends Job
                 self::MUTEX_LOCK_TIMEOUT,
                 ErrorCode::BAD_REQUEST_ANOTHER_OPERATION_IN_PROGRESS
             );
-            // delete job on successful processing
-            $this->delete();
+
+            if ($this->retryEnabled === true)
+            {
+                // delete job on successful processing
+                $this->delete();
+            }
         }
         catch (\Throwable $ex)
         {
@@ -120,8 +126,16 @@ class LedgerStatus extends Job
                     'transactor_event'  => $this->transactorEvent,
                     'ledger_request'    => $this->ledgerRequest
                 ]);
-            // retry job on failure
-            $this->checkRetry($ex);
+
+            if ($this->retryEnabled === true)
+            {
+                // retry job on failure
+                $this->checkRetry($ex);
+            }
+            else
+            {
+                throw $ex;
+            }
         }
     }
